new configs, dual database : one accessible from code and one from console.

+++ src/main/java/com/mborgraeve/arduinoserver/repository/InstructionRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.mborgraeve.arduinoserver.models.Instruction;
 
-public interface InstructionRepository extends CrudRepository<Instruction, Long> {
-	Instruction findById(Long id);
+
+public interface InstructionRepository extends CrudRepository<Instruction, Integer>{
+	Instruction findById(int id);
+
 }
